requirements.txt
easydict
einops
numpy
pandas
Pillow
torch
torchvision
